update dec w benchmark testing
